# backend/openapi.yml
openapi: 3.0.3
info:
  title: CapeConnect API
  version: 1.1.0
  description: >
    Live backend contract for the current CapeConnect static frontend and
    supporting admin tools. Authentication uses opaque bearer session tokens,
    not JWTs.
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    PaymentStatus:
      type: object
      properties:
        id:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, CANCELLED]
        description:
          type: string
        reference:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
paths:
...
  /api/payments/topup/initiate:
    post:
      tags: [Payments]
      summary: Create a PayFast wallet top-up payment session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: integer
                  description: Amount in cents
      responses:
        '200':
          description: PayFast payment form data
  /api/payments/payfast/webhook:
    post:
      tags: [Payments]
      summary: PayFast server-to-server webhook
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Validation rejected
  /api/payments/status/{paymentId}:
    get:
      tags: [Payments]
      summary: Get one payment status
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'
  /api/payments/history:
    get:
      tags: [Payments]
      summary: Get payment history for current user
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Payment history
...